Revert "Hack around HMR issues using a "data fetcher + internal" component approach"
This reverts commit 57fdd2b74d9bfe1e99534e41e5f8cd6098b47317.
